pyproject.toml
[project]
name = "cibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diff-cover>=9.2.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "msgspec>=0.19.0",
    "packaging>=24.2",
    "pydantic-settings>=2.8.1",
    "pygithub>=2.6.1",
    "typer>=0.15.2",
]


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cibot"]


[tool.pyright]
include = ["src"]
typeCheckingMode = "standard"


[dependency-groups]
dev = [
    "ruff>=0.9.10",
]
